fix an issue in AlertReward and rename the files in the Reward module - starting with lower case
