+++ grid2op/Reward/__init__.py
@@ -28,29 +28,29 @@ __all__ = [
     "_AssistantCostScore"
 ]
 
-from grid2op.Reward.ConstantReward import ConstantReward
-from grid2op.Reward.EconomicReward import EconomicReward
-from grid2op.Reward.FlatReward import FlatReward
-from grid2op.Reward.IncreasingFlatReward import IncreasingFlatReward
-from grid2op.Reward.L2RPNReward import L2RPNReward
-from grid2op.Reward.RedispReward import RedispReward
-from grid2op.Reward.BridgeReward import BridgeReward
-from grid2op.Reward.CloseToOverflowReward import CloseToOverflowReward
-from grid2op.Reward.DistanceReward import DistanceReward
-from grid2op.Reward.GameplayReward import GameplayReward
-from grid2op.Reward.LinesReconnectedReward import LinesReconnectedReward
-from grid2op.Reward.LinesCapacityReward import LinesCapacityReward
-from grid2op.Reward.CombinedReward import CombinedReward
-from grid2op.Reward.CombinedScaledReward import CombinedScaledReward
-from grid2op.Reward.RewardHelper import RewardHelper
-from grid2op.Reward.BaseReward import BaseReward
-from grid2op.Reward.L2RPNSandBoxScore import L2RPNSandBoxScore
-from grid2op.Reward.EpisodeDurationReward import EpisodeDurationReward
-from grid2op.Reward.AlarmReward import AlarmReward
-from grid2op.Reward._AlarmScore import _AlarmScore
+from grid2op.Reward.constantReward import ConstantReward
+from grid2op.Reward.economicReward import EconomicReward
+from grid2op.Reward.flatReward import FlatReward
+from grid2op.Reward.increasingFlatReward import IncreasingFlatReward
+from grid2op.Reward.l2RPNReward import L2RPNReward
+from grid2op.Reward.redispReward import RedispReward
+from grid2op.Reward.bridgeReward import BridgeReward
+from grid2op.Reward.closeToOverflowReward import CloseToOverflowReward
+from grid2op.Reward.distanceReward import DistanceReward
+from grid2op.Reward.gameplayReward import GameplayReward
+from grid2op.Reward.linesReconnectedReward import LinesReconnectedReward
+from grid2op.Reward.linesCapacityReward import LinesCapacityReward
+from grid2op.Reward.combinedReward import CombinedReward
+from grid2op.Reward.combinedScaledReward import CombinedScaledReward
+from grid2op.Reward.rewardHelper import RewardHelper
+from grid2op.Reward.baseReward import BaseReward
+from grid2op.Reward.l2RPNSandBoxScore import L2RPNSandBoxScore
+from grid2op.Reward.episodeDurationReward import EpisodeDurationReward
+from grid2op.Reward.alarmReward import AlarmReward
+from grid2op.Reward._alarmScore import _AlarmScore
 from grid2op.Reward.n1Reward import N1Reward
 from grid2op.Reward.l2rpn_wcci2022_scorefun import L2RPNWCCI2022ScoreFun
-from grid2op.Reward.AlertReward import AlertReward
+from grid2op.Reward.alertReward import AlertReward
 from grid2op.Reward._newRenewableSourcesUsageScore import _NewRenewableSourcesUsageScore
 from grid2op.Reward._assistantScore import _AssistantConfidenceScore, _AssistantCostScore
