+++ grid2op/Reward/bridgeReward.py
@@ -9,7 +9,7 @@
 import numpy as np
 import networkx as nx
 
-from grid2op.Reward.BaseReward import BaseReward
+from grid2op.Reward.baseReward import BaseReward
 from grid2op.dtypes import dt_float
