+++ grid2op/Reward/l2RPNSandBoxScore.py
@@ -8,7 +8,7 @@
 
 import numpy as np
 
-from grid2op.Reward.BaseReward import BaseReward
+from grid2op.Reward.baseReward import BaseReward
 from grid2op.dtypes import dt_float
