+++ grid2op/Reward/redispReward.py
@@ -12,7 +12,7 @@ import numpy as np
 import re
 from grid2op._glop_platform_info import _IS_WINDOWS, _IS_LINUX, _IS_MACOS
 from grid2op.Exceptions import Grid2OpException
-from grid2op.Reward.BaseReward import BaseReward
+from grid2op.Reward.baseReward import BaseReward
 from grid2op.dtypes import dt_float
